docs: Add guidance on installing the programming env
This commit updates the README file to include information on
installing R and RStudio and the necessary packages within R
for plotting, manipulating data, summarizing data and putting
the R markdown file into HTML format.  I also provided links
to the documentation on the packages I use in the R markdown
file.  I did this so users would be able to set up their
programming environment and have the necessary packages to
make contributions during the project.