feat: Add lines to obtain and set working directory
This commit adds lines for users to obtain and set the correct
working directory to access the csv files.  Users will need
to add in the working directory they have cloned the project
to inside the parentheses in the setwd() command.  I also
added a Reference for the commands I inserted.