feat: Add second Resource section for R chuncks
This commit adds a second Resource section for R chuncks and places
all references for code inside of R chuncks in that Resource
section of the Rmd file.  I did this so that those references wouldn't
be visible in the original Resources list.  The original Resource list
still  remains but only contains References for paragraphs outside the
R  chunks. This way, a user who wants to knit the Rmd file into an HTML
file won't have References visible that aren't visible in the document
because the R chucks are hidden. I also re-numbered the references in
both Resource sections accordingly and in the corresponding code and
paragraphs.