feat: Edit link for Reference 3 in RMD file
This commit edits the link for Reference 3 in the RMD file.  I had
originally separated the link to get the line to be less than 80
characters but this breaks the link apart when knitting the RMD file
to an HTML file so I decided just to let the line be greater than 80
characters.