feat: Edit paragraph analyzing credit score histogram
This commit edits the paragraph analyzing the credit score
histogram.  I may not look at the percentage of users below the
drop off who are in certain categories of loan status but I can
still see if there is a relationship between the two variables
using a faceted group of box plots.