feat: Add false values for warning/message to R chunks
This commit adds false values for warning and message to the R
chunks in the RMD file.  I did this so that warnings and messages
wouldn't show up in the knitted HTML file.  I also brought the
character length to be less than 80 characters as I saw that issue
when adding the false values for warning and message.