feat: Edit paragraph analyzing employment status
This commit edits the paragraph analyzing the employment status
proportions bar chart.  I did this because in the original
paragraph I reference an "obvious relationship".  If the
proportions were equal, there wouldn't necessarily be an obvious
relationship but more a pointer to a relationship.  I would still
need to verify this with a bi-variate chart.