feat: Add comment for grid.arrange in conditional chunk
This commit adds a comment for the grid.arrange command in the
R chunk for which I look at loan status under the best and worst
possible combination of Prosper score and rating.  I also add
a comment on the summary statements in the R chunk for which I
look at days' delinquent under the best and worst combination of
Prosper score and rating.