feat: Edit analysis of loan status versus prosper score
This commit edits the analysis of loan status versus prosper
score.  I did this because the initial wording made it sound as
if I was saying that prosper scores of 7-11 were the greatest
in the Completed category (they don't represent a majority for
this category but the percentage is greatest amongst all the four
loan status categories I care about).