feat: Edit way I filtered data in head command
This commit edits the way I filtered the data in the head command
to find the first few rows for a credit score range lower of 0
and the first few rows for an employment duration greater than 400
months.  I did this to make the code more readable and will do
it this way going forward.