feat: Remove NA's from debt-to-income-ratio in Rmd
This commit remove's the NA's from the debt to income ratio
histogram in the univariate analysis.  I did this as a way to
clean the data.  Note that the histograms don't change because R
automatically removes the NA's anyway when plotting.  I also
removed NA's from the summary and from the first few rows with
debt-to-income-ratio greater than 5.