feat: Edit y axis limit/breaks in days delinquent comp.
This commit edits the y axis limits and breaks so that they are
consistent between days' delinquent with the best possible
combination of prosper score/rating and that with the worst possible
combination.  I did this so that readers/users can easily
compare the two charts.  I also added a Reference for how I did
this (scale_y_continuous).