feat: Remove NA's from credit score vs loan Stat.
This commit removes NA's from the credit score versus loan Stat.
box plot.  I also removed the NA's from the summary of credit
score grouped by loan Stat.  I did this so that the NA's would
not be visible in the summary (note that the box plots wont
change because R automatically removes the NA's when graphing
box plots anyway).