feat: Remove NA's from debt to income vs loan Stat.
This commit remove's NA's from the debt to income ratio versus
loan Stat. box plot.  I also removed the NA's from the summary of
debt to income ratio grouped by loan Stat.  I did this so that the
NA's would not be visible in the summary (note that the box plots
wont change because R automatically removes the NA's when graphing
box plots anyway.)