feat: Add note about transparency in analysis
This commit adds a note about the transparency and the color
of the points in the legend in the graph of debt to income ratio
versus credit score filled by income range.  I did this so that
users would know that the color in the legend now represents 10
points instead of 1.