feat: Edit debt to income ratio threshold in analysis
This commit edits the debt to income ratio threshold in the
analysis to be 0.2.  I did this because this was closer to
the medians for each of the loan statuses in the box plot from
the univariate case.  For each loan status, around an equivalent
number of borrowers are below this threshold as above.