feat:  Add Reflection section and Reflection
This commit adds a Reflection section and the reflection to the
RMD file.  I added this Reflection as a way to recap what I found
during the analysis.  I also added it to let users know what
struggles I had during the analysis,  what I thought went well,
which results were expected and which results were surprising.  I
also gave a few possibilities for further exploration in the
future.